Reorganize entities. Make scripts work from any directory.
