+++ packages/pipeline/src/ormconfig.ts
@@ -1,11 +1,31 @@
 import { ConnectionOptions } from 'typeorm';
 
+import {
+    Block,
+    ExchangeCancelEvent,
+    ExchangeCancelUpToEvent,
+    ExchangeFillEvent,
+    Relayer,
+    SraOrder,
+    Transaction,
+} from './entities';
+
+const entities = [
+    Block,
+    ExchangeCancelEvent,
+    ExchangeCancelUpToEvent,
+    ExchangeFillEvent,
+    Relayer,
+    SraOrder,
+    Transaction,
+];
+
 export const testConfig: ConnectionOptions = {
     type: 'sqlite',
     database: 'database.sqlite',
     synchronize: true,
     logging: false,
-    entities: ['./lib/src/entities/**/*.js'],
+    entities,
     migrations: ['./lib/src/migrations/**/*.js'],
     cli: {
         entitiesDir: 'lib/src/entities',
@@ -18,7 +38,7 @@ export const deployConfig: ConnectionOptions = {
     url: process.env.ZEROEX_DATA_PIPELINE_DB_URL,
     synchronize: true,
     logging: false,
-    entities: ['./lib/src/entities/**/*.js'],
+    entities,
     migrations: ['./lib/src/migrations/**/*.js'],
     cli: {
         entitiesDir: 'lib/src/entities',
