+++ packages/pipeline/src/scripts/pull_missing_events.ts
@@ -5,7 +5,7 @@ import 'reflect-metadata';
 import { Connection, createConnection, Repository } from 'typeorm';
 
 import { ExchangeEventsSource } from '../data_sources/contract-wrappers/exchange_events';
-import { ExchangeFillEvent } from '../entities/ExchangeFillEvent';
+import { ExchangeFillEvent } from '../entities';
 import { deployConfig } from '../ormconfig';
 import { parseExchangeEvents } from '../parsers/events';
